requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
blis==0.7.9
boto3==1.26.126
botocore==1.29.126
bravado==11.0.3
bravado-core==5.17.1
Brotli==1.0.9
cachetools==5.3.0
captum==0.6.0
catalogue==2.0.8
catalyst==22.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comet-ml==3.33.0
comm==0.1.3
confection==0.0.4
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
docker-pycreds==0.4.0
dulwich==0.21.5
email-validator==2.0.0.post2
everett==3.1.0
executing==1.2.0
fastai==2.7.12
fastapi==0.95.2
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.0
Flask==2.3.2
Flask-Compress==1.13
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.14.1
hydra-slayer==0.4.1
idna==3.4
imageio==2.28.1
ipykernel==6.22.0
ipython==8.13.2
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kornia==0.6.12
langcodes==3.3.0
lazy_loader==0.2
lightning-utilities==0.8.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
motor==3.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
neptune-client==1.1.1
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.8.12
packaging==23.1
pandas==2.0.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pathtools==0.1.2
pathy==0.10.1
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydicom==2.3.1
Pygments==2.15.1
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-ignite==0.4.12
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
swagger-spec-validator==3.0.3
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.8.21.dev0
tokenizers==0.13.3
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
wandb==0.15.1
wasabi==1.1.1
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wrapt==1.15.0
wurlitzer==3.0.3
yarl==1.9.2
